dev-service/requirements.txt
Flask>=3.0
pytest>=7.0
pytest-cov>=4.0
pytest-html>=4.0
requests>=2.28
psycopg2-binary==2.9.9